+++ src/Support/Arr.lua
@@ -528,6 +528,63 @@ local Arr = {}
     end
 
     --[[--
+    Extracts a slice of a list.
+
+    As a first version of this method, it only accepts arrays and won't work for
+    associative tables.
+
+    @tparam table list The list to be sliced
+    @tparam integer offset The offset to start the slice
+    @tparam[opt='#list'] integer length The length of the slice, or nil to slice until the end
+
+    @treturn table The sliced list
+
+    @usage
+        local list = {1, 2, 3}
+        local results = library.arr:slice(list, 2, 2)
+        -- results = {2, 3}
+    ]]
+    function Arr:slice(list, offset, length)
+        if not self:isArray(list) then return list end
+
+        -- if the length is nil, it will slice until the end of the list
+        if length == nil then length = #list end
+
+        local results = {}
+        for i = offset, offset + length - 1 do
+            table.insert(results, list[i])
+        end
+        return results
+    end
+
+    --[[--
+    Truncates a list to a given size.
+
+    By a design decision, this method accepts only arrays, not associative
+    tables as it wouldn't produce the same results for the same input,
+    considering how Lua handles tables.
+
+    @tparam table list The list to be truncated
+    @tparam integer size The size to truncate the list to
+
+    @treturn table The truncated list
+
+    @usage
+        local list = {1, 2, 3}
+        local results = library.arr:truncate(list, 2)
+        -- results = {1, 2}
+    ]]
+    function Arr:truncate(list, size)
+        if size > #list then return list end
+
+        local results = {}
+        for i = 1, size do
+            table.insert(results, list[i])
+        end
+        return results
+    end
+
+    --[[--
     Calls the available unpack() method given the running environment.
 
     This method is an important helper because World of Warcraft supports