+++ documentation/docs/changelog.md
@@ -4,6 +4,12 @@ sidebar_position: 2
 
 # Changelog
 
+## yyyy.mm.dd - version 1.14.0
+
+* Add Arr:slice() to extract a portion of an array
+* Add Arr:truncate() to truncate arrays to a specific length
+* Add Str:startsWith() to check if a string starts with a specific substring
+
 ## 2024.09.11 - version 1.13.0
 
 * Add Arr:filter() to filter values in a table from a callback function